Refactor: Use `const` for `TaskContext` and other variables where applicable to improve code clarity and enforce immutability. Standardize formatting of variable declarations across the codebase.

+++ src/Compiler/Sema/Type/TypeManager.cpp
@@ -375,7 +375,7 @@ void TypeManager::buildPromoteTable()
         typeChar_,
         typeRune_};
 
-    constexpr uint32_t n = static_cast<uint32_t>(types.size());
+    constexpr uint32_t n = types.size();
 
     promoteIndex_.clear();
     promoteIndex_.reserve(n);
